integrate functions to class xxHelper for util, and xxInterceptor for interceptor.
