+++ resource/js/util/interceptor/MarkdownListUtil.js
@@ -1,17 +1,11 @@
-import { BasicInterceptor } from 'crowi-pluginkit';
-
 import * as codemirror from 'codemirror';
 
-import markdownTable from 'markdown-table';
-
 /**
- * The interceptor that reform markdown table
+ * Utility for markdown list
  */
-export default class AbortContinueMarkdownListInterceptor extends BasicInterceptor {
+class MarkdownListUtil {
 
   constructor() {
-    super();
-
     // https://github.com/codemirror/CodeMirror/blob/c7853a989c77bb9f520c9c530cbe1497856e96fc/addon/edit/continuelist.js#L14
     // https://regex101.com/r/7BN2fR/5
     this.indentAndMarkRE = /^(\s*)(>[> ]*|[*+-] \[[x ]\]\s|[*+-]\s|(\d+)([.)]))(\s*)/;
@@ -23,52 +17,6 @@ export default class AbortContinueMarkdownListInterceptor extends BasicIntercept
     this.getEol = this.getEol.bind(this);
     this.getStrFromBol = this.getStrFromBol.bind(this);
     this.getStrToEol = this.getStrToEol.bind(this);
-  }
-
-  /**
-   * @inheritdoc
-   */
-  isInterceptWhen(contextName) {
-    return (
-      contextName === 'preHandleEnter'
-    );
-  }
-
-  /**
-   * return boolean value whether processable parallel
-   */
-  isProcessableParallel() {
-    return false;
-  }
-
-  /**
-   * @inheritdoc
-   */
-  process(contextName, ...args) {
-    console.log(performance.now() + ': AbortContinueMarkdownListInterceptor.process is started');
-    const orgContext = args[0];
-    const editor = orgContext.editor;
-
-    console.log('AbortContinueMarkdownListInterceptor.process');
-
-    // get strings from current position to EOL(end of line) before break the line
-    const strToEol = this.getStrToEol(editor);
-    if (this.indentAndMarkRE.test(strToEol)) {
-      const context = Object.assign(args[0]);   // clone
-
-      console.log('AbortContinueMarkdownListInterceptor.newlineAndIndentContinueMarkdownList: abort auto indent');
-      codemirror.commands.newlineAndIndent(editor);
-      // replace the line with strToEol (abort auto indent)
-      editor.getDoc().replaceRange(strToEol, this.getBol(editor), this.getEol(editor));
-
-      // report to manager that handling was done
-      context.handlers.push(this.className);
-    }
-
-    console.log(performance.now() + ': AbortContinueMarkdownListInterceptor.process is finished');
-
-    // resolve
-    return Promise.resolve(orgContext);
   }
 
   /**
@@ -196,3 +144,8 @@ export default class AbortContinueMarkdownListInterceptor extends BasicIntercept
     return editor.getDoc().getRange(curPos, this.getEol(editor));
   }
 }
+
+// singleton pattern
+const instance = new MarkdownListUtil();
+Object.freeze(instance);
+export default instance;
